The validate_gait_in_directory function was previously unused but now it is integrated in the validate_versions_in_directory function.
Although the file path is already checked with every subgait this function is still nice to integrate. This function makes sure that the gait name exists before the subgaits are checked. So even if a (wrong)gait name with or without subgaits is present in the gait version map the program will recognize a faulty value.