Created a service configuration between the data collector and the balance_gait.py
In the data collector the updated center of mass is received with both position and velocity. Using the new inverted pendulum the future centor of mass is calculated which is used to calculate the capture point position. This is send over the client service setup with the new CapturePointPose.srv.